--- app/src/main/java/com/mashup/hongsam/di/ApplicationComponent.kt
@@ -1,16 +1,0 @@
-package com.mashup.hongsam.di
-
-import android.content.Context
-import dagger.BindsInstance
-import dagger.Component
-import javax.inject.Singleton
-
-@Singleton
-@Component(modules = [CoroutineModule::class, NetworkModule::class])
-interface ApplicationComponent {
-
-    @Component.Factory
-    interface Factory {
-        fun create(@BindsInstance context: Context): ApplicationComponent
-    }
-}
