--- app/src/main/java/com/mashup/hongsam/di/CoroutineModule.kt
@@ -1,38 +1,0 @@
-package com.mashup.hongsam.di
-
-import dagger.Module
-import dagger.Provides
-import kotlinx.coroutines.CoroutineDispatcher
-import kotlinx.coroutines.Dispatchers
-import javax.inject.Qualifier
-import javax.inject.Singleton
-
-@Module
-class CoroutineModule {
-    @Singleton
-    @Provides
-    @IoDispatcher
-    fun providesIoDispatcher(): CoroutineDispatcher = Dispatchers.IO
-
-    @Singleton
-    @Provides
-    @MainDispatcher
-    fun providesMainDispatcher(): CoroutineDispatcher = Dispatchers.Main
-
-    @Singleton
-    @Provides
-    @DefaultDispatcher
-    fun providesDefaultDispatcher(): CoroutineDispatcher = Dispatchers.Default
-}
-
-@Retention(AnnotationRetention.RUNTIME)
-@Qualifier
-annotation class DefaultDispatcher
-
-@Retention(AnnotationRetention.RUNTIME)
-@Qualifier
-annotation class IoDispatcher
-
-@Retention(AnnotationRetention.RUNTIME)
-@Qualifier
-annotation class MainDispatcher
