+++ app/src/main/java/com/mashup/twotoo/di/ApplicationComponent.kt
@@ -1,0 +1,16 @@
+package com.mashup.twotoo.di
+
+import android.content.Context
+import dagger.BindsInstance
+import dagger.Component
+import javax.inject.Singleton
+
+@Singleton
+@Component(modules = [CoroutineModule::class, NetworkModule::class])
+interface ApplicationComponent {
+
+    @Component.Factory
+    interface Factory {
+        fun create(@BindsInstance context: Context): ApplicationComponent
+    }
+}
