--- data/src/main/java/com/mashup/hongsam/di/NetworkModule.kt
@@ -1,52 +1,0 @@
-package com.mashup.hongsam.di
-
-import com.mashup.hongsam.data.BuildConfig
-import dagger.Module
-import dagger.Provides
-import okhttp3.OkHttpClient
-import okhttp3.logging.HttpLoggingInterceptor
-import retrofit2.Retrofit
-import retrofit2.converter.moshi.MoshiConverterFactory
-import javax.inject.Singleton
-
-@Module
-class NetworkModule {
-
-
-
-    @Provides
-    @Singleton
-    fun provideInterceptor(): HttpLoggingInterceptor {
-        return HttpLoggingInterceptor().setLevel(HttpLoggingInterceptor.Level.BODY)
-    }
-
-    @Provides
-    @Singleton
-    fun provideOkHttpClient(httpLoggingInterceptor: HttpLoggingInterceptor): OkHttpClient {
-        val okHttpClientBuilder = OkHttpClient.Builder()
-        if (BuildConfig.DEBUG) {
-            okHttpClientBuilder.addInterceptor(httpLoggingInterceptor)
-        }
-        return okHttpClientBuilder.build()
-    }
-
-    @Provides
-    @Singleton
-    fun provideMoshiConverterFactory(): MoshiConverterFactory {
-        return MoshiConverterFactory.create()
-    }
-
-    @Provides
-    @Singleton
-    fun provideRetrofit(client: OkHttpClient, moshiConverterFactory: MoshiConverterFactory): Retrofit {
-        return Retrofit.Builder().baseUrl(URL)
-            .client(client).addConverterFactory(
-                moshiConverterFactory,
-            ).build()
-    }
-
-    companion object {
-        const val URL = ""
-    }
-
-}
