+++ presenter/src/main/AndroidManifest.xml
@@ -1,10 +1,10 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.mashup.hongsam">
+    package="com.mashup.twotoo">
 
     <application>
         <activity
-            android:name="com.mashup.hongsam.presenter.MainActivity"
+            android:name="com.mashup.twotoo.presenter.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
