--- presenter/src/main/java/com/mashup/hongsam/presenter/di/ViewModelFactory.kt
@@ -1,17 +1,0 @@
-package com.mashup.hongsam.presenter.di
-
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.ViewModelProvider
-import javax.inject.Inject
-import javax.inject.Provider
-import javax.inject.Singleton
-
-@Suppress("UNCHECKED_CAST")
-@Singleton
-class ViewModelFactory @Inject constructor(
-    private val viewModelMap: Map<Class<out ViewModel>, @JvmSuppressWildcards Provider<ViewModel>>,
-) : ViewModelProvider.Factory {
-    override fun <T : ViewModel> create(modelClass: Class<T>): T {
-        return viewModelMap[modelClass]?.get() as T
-    }
-}
