--- presenter/src/main/java/com/mashup/hongsam/presenter/di/ViewModelFactoryModule.kt
@@ -1,12 +1,0 @@
-package com.mashup.hongsam.presenter.di
-
-import androidx.lifecycle.ViewModelProvider
-import dagger.Binds
-import dagger.Module
-
-@Module
-abstract class ViewModelFactoryModule {
-
-    @Binds
-    abstract fun bindViewModelFactory(viewModelFactory: ViewModelFactory): ViewModelProvider.Factory
-}
