--- presenter/src/main/java/com/mashup/hongsam/presenter/di/ViewModelKey.kt
@@ -1,9 +1,0 @@
-package com.mashup.hongsam.presenter.di
-
-import androidx.lifecycle.ViewModel
-import dagger.MapKey
-import kotlin.reflect.KClass
-
-@MapKey
-@Target(AnnotationTarget.FUNCTION, AnnotationTarget.PROPERTY_GETTER, AnnotationTarget.PROPERTY_SETTER)
-annotation class ViewModelKey(val value: KClass<out ViewModel>)
