--- presenter/src/main/java/com/mashup/hongsam/presenter/di/ViewModelModule.kt
@@ -1,6 +1,0 @@
-package com.mashup.hongsam.presenter.di
-
-import dagger.Module
-
-@Module
-class ViewModelModule
