+++ presenter/src/main/java/com/mashup/twotoo/presenter/di/ViewModelFactoryModule.kt
@@ -1,0 +1,12 @@
+package com.mashup.twotoo.presenter.di
+
+import androidx.lifecycle.ViewModelProvider
+import dagger.Binds
+import dagger.Module
+
+@Module
+abstract class ViewModelFactoryModule {
+
+    @Binds
+    abstract fun bindViewModelFactory(viewModelFactory: ViewModelFactory): ViewModelProvider.Factory
+}
