+++ presenter/src/main/java/com/mashup/twotoo/presenter/di/ViewModelKey.kt
@@ -1,0 +1,9 @@
+package com.mashup.twotoo.presenter.di
+
+import androidx.lifecycle.ViewModel
+import dagger.MapKey
+import kotlin.reflect.KClass
+
+@MapKey
+@Target(AnnotationTarget.FUNCTION, AnnotationTarget.PROPERTY_GETTER, AnnotationTarget.PROPERTY_SETTER)
+annotation class ViewModelKey(val value: KClass<out ViewModel>)
