+++ presenter/src/main/java/com/mashup/twotoo/presenter/di/ViewModelModule.kt
@@ -1,0 +1,6 @@
+package com.mashup.twotoo.presenter.di
+
+import dagger.Module
+
+@Module
+class ViewModelModule
